Finished magic missile (beta)
Why should I put a description for this? I find it quite pointless for a description to be here, all you need to know is in the title. It's really quite simple; I just added a few VFX and made it spin. That's pretty much it. Do you really need this? Was this worth your time? I hope you spent way too long reading this, as it is a time wasting description and serves almost zero purpose and stuff. Now go back to work peasant