Fixed a couple bugs with weapon select and worked on consumables
Fixed a bug with selecting a weapon while an animation was playing. basically it just checks if the animation is playing based on a boolean that gets set on when they animation starts and put off when it's done. I also added Imp in a Bottle to the weapon enum and made an actor, not a lot done there quite yet though. I also made it so you equip the weapons with 1 and 2 but strictly if the spell wheel is closed.